Changed: #9, #31, Outsourced into many classes
There is now an extra file for each table which is to be created.
It is now "class based", which increases the comfort and overview of it.
Btw, I think I know now what .pyc means.. "python class".. So there is no need to delete them because they are "necessary"